+++ app/Actions/Employee/DeleteEmployeeAction.php
@@ -2,13 +2,18 @@
 
 namespace App\Actions\Employee;
 
+use App\Exceptions\NotFoundException;
 use App\Models\User;
 
 class DeleteEmployeeAction
 {
     public function execute(int $id): void
     {
-        $user = User::with('employee')->findOrFail($id);
+        $user = User::with('employee')->find($id);
+
+        if (! $user) {
+            throw new NotFoundException('Employee not found');
+        }
 
         if ($user->employee) {
             $user->employee->delete();
